Create a heatmap of correlation between receptors and transcription factors

Creates a heatmap of correlation values between receptors and transcription factors.

Usage

cor_heatmap(
  dom,
  bool = TRUE,
  bool_thresh = 0.15,
  title = TRUE,
  feats = NULL,
  recs = NULL,
  mark_connections = FALSE,
  ...
)

Arguments

dom

A domino object with network built (build_domino)

bool

A boolean indicating whether the heatmap should be continuous or boolean. If boolean then bool_thresh will be used to determine how to define activity as positive or negative.

bool_thresh

A numeric indicating the threshold separating 'on' or 'off' for feature activity if making a boolean heatmap.

title

Either a string to use as the title or a boolean describing whether to include a title. In order to pass the 'main' parameter to NMF::aheatmap you must set title to FALSE.

feats

Either a vector of features to include in the heatmap or 'all' for all features. If left NULL then the features selected for the signaling network will be shown.

recs

Either a vector of receptors to include in the heatmap or 'all' for all receptors. If left NULL then the receptors selected in the signaling network connected to the features plotted will be shown.

mark_connections

A boolean indicating whether to add an 'x' in cells where there is a connected receptor or TF. Default FALSE.

...

Other parameters to pass to NMF::aheatmap. Note that to use the 'main' parameter of NMF::aheatmap you must set title = FALSE and to use 'annCol' or 'annColors' ann_cols must be FALSE.