Create a domino object and prepare it for network construction

This function reads in a receptor ligand signaling database, cell level features of some kind (ie. output from pySCENIC), z-scored single cell data, and cluster id for single cell data, calculates a correlation matrix between receptors and other features (this is transcription factor module scores if using pySCENIC), and finds features enriched by cluster. It will return a domino object prepared for build_domino(), which will calculate a signaling network.

Usage

create_domino(
  rl_map,
  features,
  ser = NULL,
  counts = NULL,
  z_scores = NULL,
  clusters = NULL,
  use_clusters = TRUE,
  tf_targets = NULL,
  verbose = TRUE,
  use_complexes = TRUE,
  rec_min_thresh = 0.025,
  remove_rec_dropout = TRUE,
  tf_selection_method = "clusters",
  tf_variance_quantile = 0.5
)

Arguments

rl_map

Data frame where each row describes a receptor-ligand interaction with required columns gene_A & gene_B including the gene names for the receptor and ligand and type_A & type_B annotating if genes A and B are a ligand (L) or receptor (R)

features

Either a path to a csv containing cell level features of interest (ie. the auc matrix from pySCENIC) or named matrix with cells as columns and features as rows.

ser

Seurat object containing scaled RNA expression data in the RNA assay slot and cluster identity. Either a ser object OR z_scores and clusters must be provided. If ser is present z_scores and clusters will be ignored.

counts

Counts matrix for the data. If a Seurat object is provided this will be ignored. This is only used to threshold receptors on dropout.

z_scores

Matrix containing z-scored expression data for all cells with cells as columns and features as rows. Either z_scores and clusters must be provided OR a ser object. If ser is present z_scores and clusters will be ignored.

clusters

Named factor containing cell cluster with names as cells. Either clusters and z_scores OR ser must be provided. If ser is present z_scores and clusters will be ignored.

use_clusters

Boolean indicating whether to use the clusters from a Seurat object. If a Seurat object is not provided then this parameter is ignored.

tf_targets

Optional. A list where names are transcription factors and the stored values are character vectors of genes in the transcription factor's regulon.

verbose

Boolean indicating whether or not to print progress during computation.

use_complexes

Boolean indicating whether you wish to use receptor/ligand complexes in the receptor ligand signaling database. If FALSE, receptor/ligand pairs where either functions as a protein complex will not be considered when constructing the signaling network.

rec_min_thresh

Minimum expression level of receptors by cell. Default is 0.025 or 2.5 percent of all cells in the data set. This is important when calculating correlation to connect receptors to transcription activation. If this threshold is too low then correlation calculations will proceed with very few cells with non-zero expression.

remove_rec_dropout

Whether to remove receptors with 0 expression counts when calculating correlations. This can reduce false positive correlation calculations when receptors have high dropout rates.

tf_selection_method

Selection of which method to target transcription factors. If 'clusters' then differential expression for clusters will be calculated. If 'variable' then the most variable transcription factors will be selected. If 'all' then all transcription factors in the feature matrix will be used. Default is 'clusters'. Note that if you wish to use clusters for intercellular signaling downstream to MUST choose clusters.

tf_variance_quantile

What proportion of variable features to take if using variance to threshold features. Default is 0.5. Higher numbers will keep more features. Ignored if tf_selection_method is not 'variable'

Value

A domino object

Examples

pbmc_dom_tiny_all <- pbmc_dom_tiny <- create_domino(
    rl_map = domino2:::rl_map_tiny, features = domino2:::auc_tiny,
    counts = domino2:::RNA_count_tiny, z_scores = domino2:::RNA_zscore_tiny,
    clusters = domino2:::clusters_tiny, tf_targets = domino2:::regulon_list_tiny,
    use_clusters = FALSE, use_complexes = FALSE,
    rec_min_thresh = 0.1, remove_rec_dropout = TRUE,
    tf_selection_method = "all"
)
#> Reading in and processing signaling database
#> Database provided from source: CellPhoneDB
#> Getting z_scores, clusters, and counts
#> Calculating correlations
#> 1 of 3
#> 2 of 3
#> 3 of 3
#> FAS has component genes that did not pass testing parameters
#> CD22 has component genes that did not pass testing parameters
#> IGF1R has component genes that did not pass testing parameters

pbmc_dom_tiny_clustered <- create_domino(
    rl_map = domino2:::rl_map_tiny, features = domino2:::auc_tiny,
    counts = domino2:::RNA_count_tiny, z_scores = domino2:::RNA_zscore_tiny,
    clusters = domino2:::clusters_tiny, tf_targets = domino2:::regulon_list_tiny,
    use_clusters = TRUE, use_complexes = TRUE, remove_rec_dropout = FALSE
)
#> Reading in and processing signaling database
#> Database provided from source: CellPhoneDB
#> Getting z_scores, clusters, and counts
#> Calculating feature enrichment by cluster
#> 1 of 3
#> 2 of 3
#> 3 of 3
#> Calculating correlations
#> 1 of 3
#> Warning: Cannot compute exact p-value with ties
#> Warning: Cannot compute exact p-value with ties
#> Warning: Cannot compute exact p-value with ties
#> Warning: Cannot compute exact p-value with ties
#> Warning: Cannot compute exact p-value with ties
#> 2 of 3
#> Warning: Cannot compute exact p-value with ties
#> Warning: Cannot compute exact p-value with ties
#> Warning: Cannot compute exact p-value with ties
#> Warning: Cannot compute exact p-value with ties
#> Warning: Cannot compute exact p-value with ties
#> 3 of 3
#> Warning: Cannot compute exact p-value with ties
#> Warning: Cannot compute exact p-value with ties
#> Warning: Cannot compute exact p-value with ties
#> Warning: Cannot compute exact p-value with ties
#> Warning: Cannot compute exact p-value with ties
#> IGF1R has component genes that did not pass testing parameters